+++ NebulaNetwork/PacketProcessors/Session/StartGameMessageProcessor.cs
@@ -64,6 +64,7 @@ namespace NebulaNetwork.PacketProcessors.Session
                 ((LocalPlayer)Multiplayer.Session.LocalPlayer).SetPlayerData(packet.LocalPlayerData, true);
 
                 UIRoot.instance.uiGame.planetDetail.gameObject.SetActive(false);
+                Multiplayer.Session.IsInLobby = false;
 
                 GameDesc gameDesc = UIRoot.instance.galaxySelect.gameDesc;
                 gameDesc.SetForNewGame(gameDesc.galaxyAlgo, gameDesc.galaxySeed, gameDesc.starCount, 1, gameDesc.resourceMultiplier);